Kibble/requirements.txt
amqp==2.1.4
appdirs==1.4.3
beautifulsoup4==4.6.0
billiard==3.5.0.2
boto3==1.4.8
celery==4.1.0
chardet==2.2.1
coverage==4.4.2
cryptography==2.1.4
Django==1.10.5
django-allauth==0.30.0
django-constance==2.0.0
django-jsonfield==1.0.1
django-model-utils==2.2
django-queryinspect==1.0.0
djangorestframework==3.5.3
django-storages==1.6.5
django-watchman==0.15.0
dj-database-url==0.4.1
dropbox==7.3.1
enum34==1.1.6
enum-compat==0.0.2
eventlet==0.20.1
Fabric==1.14.0
flake8==2.6.2
funcsigs==1.0.2
greenlet==0.4.12
gunicorn==19.6.0
inflect==0.2.5
img2pdf==0.2.4
itsdangerous==0.24
kombu==4.0.2
lxml==3.7.3
marisa-trie==0.7.4
mccabe==0.5.3
mock==2.0.0
model-mommy==1.3.2
numpy==1.12.0
oauth2client==4.1.2
oauthlib==2.0.2
packaging==16.8
pbr==3.0.1
#psycopg2
pyasn1==0.4.2
pyasn1_modules==0.2.1
pycodestyle==2.0.0
pydrive==1.3.1
pyflakes==1.2.3
pyparsing==2.2.0
pyPdf==1.13
PyStemmer==1.3.0
python-dateutil==2.6.0
python-docx==0.8.6
python-dotenv==0.7.1
python-http-client==2.2.1
python-magic==0.4.13
python-mimeparse==1.6.0
python-openid==2.2.5
pytz==2016.10
redis==2.10.5
requests-oauthlib==0.8.0
requests[security]==2.14.2
scipy==0.19.0
scikit-learn==0.18.1
sendgrid==3.6.3
sendgrid-django==4.0.4
six==1.10.0
#spacy==2.0.9
#vine==1.1.3
